Add preliminary types for JSON parser.
Added JSON__ValueType. This is an enumerator type which describes the type of a JSON value. We will store JSON values as a union; this will be used for identifying which field of the union to access. Types are: Invalid, Null, Boolean, String, Integer, Real, Array, Object, COUNT

Added struct type JSON__Value. This is a tagged union, which will be used to store JSON values.

Added struct type JSON__Object. This is a structure which holds a key-value pair corresponding to a JSON object. It also holds a reference to its parent, if it exists.